+++ Packages/com.thenathannator.plasticband/PlasticBand/Devices/Keytar/PS3WiiKeytar.cs
@@ -92,7 +92,8 @@ namespace PlasticBand.Devices.LowLevel
         [InputControl(name = "overdrive", layout = "Button", bit = 7)]
         public byte overdrive;
 
-        [InputControl(name = "analogPedal", layout = "Axis", format = "BIT", sizeInBits = 7)]
+        // TODO: The normalization here needs verification
+        [InputControl(name = "analogPedal", layout = "Axis", format = "BIT", sizeInBits = 7, parameters = "normalize,normalizeMin=1,normalizeMax=0,normalizeZero=1")]
         [InputControl(name = "digitalPedal", layout = "Button", bit = 7)]
         public byte pedal;